# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Nexus Chat Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /groups/{groupId}/members:
    get:
      operationId: listGroupMembers
      tags: [groups]
      summary: List members of a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroupMember"
    post:
      operationId: addGroupMember
      tags: [groups]
      summary: Add a member to a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupMemberInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupMember"
  /groups/{groupId}/members/{userId}:
    delete:
      operationId: removeGroupMember
      tags: [groups]
      summary: Remove a member from a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
  /groups/{groupId}/members/{userId}/role:
    patch:
      operationId: updateGroupMemberRole
      tags: [groups]
      summary: Update a member role in a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupRoleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupMember"

  # ─── CHANNELS ────────────────────────────────────────────────────────
  /channels:
    get:
      operationId: listChannels
      tags: [channels]
      summary: List public broadcast channels
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Channel"
    post:
      operationId: createChannel
      tags: [channels]
      summary: Create a broadcast channel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelInput"
      responses:
        "201":
          description: Channel created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
  /channels/{channelId}:
    get:
      operationId: getChannel
      tags: [channels]
      summary: Get channel details
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
  /channels/{channelId}/subscribe:
    post:
      operationId: subscribeChannel
      tags: [channels]
      summary: Subscribe to a channel
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Subscribed
  /channels/{channelId}/unsubscribe:
    post:
      operationId: unsubscribeChannel
      tags: [channels]
      summary: Unsubscribe from a channel
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Unsubscribed

  # ─── NOTIFICATIONS ───────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username, email, displayName, status, role, createdAt]
      properties:
        id:          { type: integer }
        username:    { type: string }
        email:       { type: string }
        displayName: { type: string }
        avatarUrl:   { type: ["string", "null"] }
        bio:         { type: ["string", "null"] }
        status:      { type: string, enum: [active, suspended, banned] }
        presence:    { type: string, enum: [online, away, offline] }
        role:        { type: string, enum: [user, moderator, admin] }
        isVerified:  { type: boolean }
        lastSeenAt:  { type: ["string", "null"] }
        createdAt:   { type: string }
...
    GroupMember:
      type: object
      required: [userId, role, joinedAt]
      properties:
        userId:   { type: integer }
        user:
          $ref: "#/components/schemas/User"
        role:     { type: string, enum: [owner, admin, moderator, member] }
        joinedAt: { type: string }
    GroupMemberInput:
      type: object
      required: [userId]
      properties:
        userId: { type: integer }
        role:   { type: string, enum: [admin, moderator, member] }
    GroupRoleUpdate:
      type: object
      required: [role]
      properties:
        role: { type: string, enum: [admin, moderator, member] }

    # ─── CHANNELS ────────────────────────────────────────────────────
    Channel:
      type: object
      required: [id, name, createdAt]
      properties:
        id:              { type: integer }
        name:            { type: string }
        description:     { type: ["string", "null"] }
        avatarUrl:       { type: ["string", "null"] }
        subscriberCount: { type: integer }
        isSubscribed:    { type: boolean }
        isPublic:        { type: boolean }
        ownerId:         { type: integer }
        createdAt:       { type: string }
    ChannelInput:
      type: object
      required: [name]
      properties:
        name:        { type: string, minLength: 2 }
        description: { type: string }
        avatarUrl:   { type: string }
        isPublic:    { type: boolean }

    # ─── NOTIFICATIONS ───────────────────────────────────────────────
...