# lib/api-zod/src/generated/types/channel.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nexus Chat Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Channel {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  subscriberCount?: number;
  isSubscribed?: boolean;
  isPublic?: boolean;
  ownerId?: number;
  createdAt: string;
}